Squashed commit of the following:
commit 25d9a0b219d34e91e585a1af99431370b1ac8635
Author: Prajjwal Poudel <iamprazol@gmail.com>
Date:   Mon Oct 9 08:11:58 2023 +0545

    Fix - Critical error due to wrong indentation

commit b7d6ef335d186736e01589acb9c2cd5879a9514f
Author: Salim Shrestha <salimshrestha98@gmail.com>
Date:   Mon Oct 9 07:15:27 2023 +0545

    UR-1200 Feature - Admin image uploader module (#608)

    * Enhance - User registration Conversational Forms dependency

    * Dev - Add wrapper for buttons

    * Fix - Uploaded image hide/show issue

commit baf30eb88a8bf1732ff320621c2d9e83262210b6
Author: Lekhnath Pandey <46083608+lekhnath005@users.noreply.github.com>
Date:   Fri Oct 6 12:07:59 2023 +0545

    Enhance - Display Fatal errors in logs (#611)

commit e04db7836abfbee8e99f57d6c2bcc900e866b061
Author: Milan Chaudhary <109399375+MILAN88888@users.noreply.github.com>
Date:   Fri Oct 6 10:48:36 2023 +0545

    Added - Form validaton compatible (#606)

commit f89b9e6a91ed564265697d1e70c32a92f77ae817
Author: Prajjwal Poudel <iamprazol@gmail.com>
Date:   Fri Oct 6 10:45:25 2023 +0545

    UR-1220 Tweak - Urls in readme txt

commit 443270be65c8050800fc102174eb6a05e366db25
Author: Milan Chaudhary <109399375+MILAN88888@users.noreply.github.com>
Date:   Thu Oct 5 16:58:02 2023 +0545

    Tweak -  Save the user IP on the user meta. (#609)

    * Added - User Ip in user meta

    * Added - Update user ip while updating profile

    * Removed - Unwanted code

    ---------


commit 86c8d1b51dd3149dd93b17877ce9501dc77d9e83
Author: Salim Shrestha <salimshrestha98@gmail.com>
Date:   Thu Oct 5 16:54:11 2023 +0545

    UR-749 Dev - Merge two duplicate methods on form validation (#590)

    * Dev - Merge two duplicate validation methods for update profile

    * Update sync branch

    * Dev - Add skippable fields on update validation

    * Tweak - Remove unnecessary function

commit 1a26ec2a6e79aebd2194d861d1caaa6c390d7703
Author: Ranjan Khanal <83644457+zhea-son@users.noreply.github.com>
Date:   Thu Oct 5 11:20:44 2023 +0545

    send user approval link in email to admin using send email option enable/disable (#610)

commit e17ea746d751aae4bc67eb0174e40bbcb043a8d2
Author: Prajjwal Poudel <iamprazol@gmail.com>
Date:   Thu Oct 5 11:00:05 2023 +0545

    UR-1066 Dev - Compatibility for custom user listings

commit f906257d59548c03b6c42fe9ba3351bcfa0f11f7
Author: Salim Shrestha <salimshrestha98@gmail.com>
Date:   Thu Oct 5 08:55:18 2023 +0545

    UR-1139 Refactor - Handle user status properly (#595)

    * Fix - User pending notice showing same number after denying users

    * Fix - Pending notice not showing

    * Fix - Pending filter showing denied users

    * Refactor - Get user login option from user meta

    Instead of handling user with the login option set in form settings in present, user must be handled according to the login option set during the time of registration.

    * Add - is_email_pending function

    * Tweak - Check function exists

    * Refactor - Add email pending label

    * Dev - Update sync parameters

    * Fix - User not being denied

    * Fix - Denied filter not working

    * Remove sync code

    ---------


commit fa8a128abeadf6fe89de1284ef62be2456b8cb9e
Author: Dipu chaudhary <34309829+dipuchaudhary@users.noreply.github.com>
Date:   Thu Oct 5 08:46:40 2023 +0545

    UR-1049 profile details updated email to user (#603)

    * Feature - Send email to user after profile details update

    * Tweak - Priority order of send profile changed email

    * Fix - Privacy field value

    * Dynamic email translation added

commit a526dfee275f4d0cddb3e9912859adee46eadcd1
Author: Ranjan Khanal <83644457+zhea-son@users.noreply.github.com>
Date:   Thu Oct 5 08:46:12 2023 +0545

    Fix - Collapse issue (#607)

commit 4451cb625e2e66179deafb0f311729a056a2e812
Author: Salim Shrestha <salimshrestha98@gmail.com>
Date:   Wed Oct 4 19:07:06 2023 +0545

    UR-870 Enhance - Test captcha from settings (#601)

    * Feature - Test captcha from captcha settings

    * Add - Cloudflare turnstile

    * Enhance - Notice messages

    * Enhance - Make strings translatable

    * Fix - Hide test captcha button on changes

    * Tweak - Hide cloudflare turnstile test wrapper when Theme of cloudflare changed

    ---------


commit 84927fb1922b3437facdf195f4802b2f1867035a
Author: Milan Chaudhary <chaudharymilan996@gmail.com>
Date:   Wed Oct 4 17:07:54 2023 +0545

    Fix - Missing closing tag

commit f8f1b5f192c0c16b7ab8efc02da961040433a799
Author: Dipu chaudhary <34309829+dipuchaudhary@users.noreply.github.com>
Date:   Wed Oct 4 08:29:19 2023 +0545

    UR-1035 Compatibility custom captcha (#604)

    * Dependency of custom captcha field

    * Custom captcha field icon added

    * Unset captcha field form edit profile

    * Dev - Captcha dependency moved to pro

    * Code cleanup

    * Updated

commit 7409448777f42726b20db2a0800438a59020dc8e
Author: Milan Chaudhary <109399375+MILAN88888@users.noreply.github.com>
Date:   Wed Oct 4 08:24:07 2023 +0545

    Tweak - Redirect url as slug in login and my account shortcode (#599)

    * Tweak - Redirection even if the redirect_url is page slug

    * Added - Redirection even if we put the slug as redirect_url in login shortcode

    * Fixed - The case if no valid url and page slug then redirect to home url

    ---------


commit 995ba2b531388c82398178994008646a04004dd9
Author: deepench <elendoff11@gmail.com>
Date:   Tue Oct 3 12:20:08 2023 +0545

    Added - WPML compatibilty for validation message for duplicate

commit 4490a547e26748aeda26b4f8ed49a50e1473e562
Author: Prajjwal Poudel <iamprazol@gmail.com>
Date:   Tue Oct 3 11:40:11 2023 +0545

    Fix - Typo

commit 56bfe9cc72754b36677a334d4b52c8a8a3b14238
Author: Salim Shrestha <salimshrestha98@gmail.com>
Date:   Tue Oct 3 11:27:28 2023 +0545

    UR-576 Fix - Select All button not working in Nav Menus (#598)

    * Fix - Select All btn not working in Nav Menus

    * Fix - Translation

commit 0aee447bc96874bcc9d4accd34ebfe48bf56ecf7
Author: Milan Chaudhary <chaudharymilan996@gmail.com>
Date:   Mon Oct 2 16:05:42 2023 +0545

    Fix - Grammatical typo for unique validation

commit 67b1296f5e93a8149432f0482c70c11778ed79b4
Author: Salim Shrestha <salimshrestha98@gmail.com>
Date:   Mon Oct 2 08:00:43 2023 +0545

    UR-1059 Enhance - Links in admin bar menu (#602)

    * Enhance - Add Edit Form and New form links in admin bar menu

    * Tweak - Create new menu for UR items

    * Tweak - Translatable text wrapper missing

    ---------


commit 67ec5a49972e855cf4c30bc842dc2c4f6366c464
Author: Prajjwal Poudel <85762481+iamprazol@users.noreply.github.com>
Date:   Mon Oct 2 07:59:46 2023 +0545

    Tweak - Addons content made concrete and design to make it fit (#605)

commit d1eac1173a76a4973e71dea078682bd3a1ea0051
Author: Dipu chaudhary <34309829+dipuchaudhary@users.noreply.github.com>
Date:   Mon Oct 2 07:54:05 2023 +0545

    jQuery validation library updated to 1.19.5 (#612)

commit 92ffda7f2b593773014dc02183318da257634ed4
Author: Deependra Chaudhary <elendoff11@gmail.com>
Date:   Mon Oct 2 07:52:36 2023 +0545

    Fix - Escaping issue in redirect url (#613)

commit 0f3a77d29c9ba17a922223f915c1f5de00af1f76
Author: Milan Chaudhary <chaudharymilan996@gmail.com>
Date:   Tue Sep 26 16:52:01 2023 +0545

    Fixed - Locked field pop issue

commit 937ed20d0b9a8d2767de75ad6c720be887df9a8c
Author: Salim Shrestha <salimshrestha98@gmail.com>
Date:   Thu Sep 21 12:14:04 2023 +0545

    Fix - Send password reset not working on prevent core login

commit 1f7dffa2f25fffb8026766b2ca9e372eb30629bb
Author: Lekhnath Pandey <lekhnath005@gmail.com>
Date:   Wed Sep 13 13:19:14 2023 +0545

    Dev - Introduced new hooks to modify email header

commit f94a8bd0f2b26151f9471a083d3ea0dbf32c8c2a
Author: Lekhnath Pandey <lekhnath005@gmail.com>
Date:   Thu Sep 7 12:47:27 2023 +0545

    Fix - Admin email not been delivered

+++ includes/log-handlers/class-ur-log-handler-file.php
@@ -247,17 +247,62 @@ class UR_Log_Handler_File extends UR_Log_Handler {
 	 */
 	public function remove( $handle ) {
 		$removed = false;
-		$file    = self::get_log_file_path( $handle );
+		$logs    = $this->get_log_files();
+		$handle  = sanitize_title( $handle );
 
-		if ( $file ) {
-			if ( is_file( $file ) && is_writable( $file ) ) {
-				$this->close( $handle ); // Close first to be certain no processes keep it alive after it is unlinked.
-				$removed = unlink( $file );
+		if ( isset( $logs[ $handle ] ) && $logs[ $handle ] ) {
+			$file = realpath( trailingslashit( UR_LOG_DIR ) . $logs[ $handle ] );
+			if ( 0 === stripos( $file, realpath( trailingslashit( UR_LOG_DIR ) ) ) && is_file( $file ) && is_writable( $file ) ) { // phpcs:ignore WordPress.VIP.FileSystemWritesDisallow.file_ops_is_writable
+				$this->close( $file ); // Close first to be certain no processes keep it alive after it is unlinked.
+				$removed = unlink( $file ); // phpcs:ignore WordPress.VIP.FileSystemWritesDisallow.file_ops_unlink
 			}
 			do_action( 'user_registration_log_remove', $handle, $removed );
 		}
-
 		return $removed;
+	}
+
+	/**
+	 * Remove/delete all log files.
+	 *
+	 * @return bool
+	 */
+	public function remove_all() {
+		$removed = false;
+		$logs    = $this->get_log_files();
+
+		if ( count( $logs ) ) {
+			foreach ( $logs as $key => $log ) {
+				$file = realpath( trailingslashit( UR_LOG_DIR ) . $log );
+				if ( 0 === stripos( $file, realpath( trailingslashit( UR_LOG_DIR ) ) ) && is_file( $file ) && is_writable( $file ) ) { // phpcs:ignore WordPress.VIP.FileSystemWritesDisallow.file_ops_is_writable
+					$this->close( $file ); // Close first to be certain no processes keep it alive after it is unlinked.
+					$removed = unlink( $file ); // phpcs:ignore WordPress.VIP.FileSystemWritesDisallow.file_ops_unlink
+				}
+			}
+		}
+		return $removed;
+	}
+
+	/**
+	 * Get all log files in the log directory.
+	 *
+	 * @since 1.6.2
+	 * @return array
+	 */
+	public static function get_log_files() {
+		$files  = @scandir( UR_LOG_DIR ); // @codingStandardsIgnoreLine.
+		$result = array();
+
+		if ( ! empty( $files ) ) {
+			foreach ( $files as $key => $value ) {
+				if ( ! in_array( $value, array( '.', '..' ), true ) ) {
+					if ( ! is_dir( $value ) && strstr( $value, '.log' ) ) {
+						$result[ sanitize_title( $value ) ] = $value;
+					}
+				}
+			}
+		}
+
+		return $result;
 	}
 
 	/**
@@ -301,7 +346,7 @@ class UR_Log_Handler_File extends UR_Log_Handler {
 	 * @param string $handle Log handle
 	 */
 	protected function log_rotate( $handle ) {
-		for ( $i = 8; $i >= 0; $i -- ) {
+		for ( $i = 8; $i >= 0; $i-- ) {
 			$this->increment_log_infix( $handle, $i );
 		}
 		$this->increment_log_infix( $handle );
@@ -336,7 +381,6 @@ class UR_Log_Handler_File extends UR_Log_Handler {
 		} else {
 			return false;
 		}
-
 	}
 
 	/**
@@ -377,5 +421,4 @@ class UR_Log_Handler_File extends UR_Log_Handler {
 			$this->add( $log['entry'], $log['handle'] );
 		}
 	}
-
 }
